book_catalog/schemas/usuario.py
from pydantic import BaseModel, EmailStr


class UserBase(BaseModel):
    username: str
    email: EmailStr


class UserCreate(UserBase):
    password: str


class UserPublic(UserBase):
    id: int


class UserUpdate(UserBase):
    password: str


class UserDeleted(BaseModel):
    message: str
