pyproject.toml
[project]
name = "book-catalog"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.118.0,<0.119.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.13.3,<0.14.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "testcontainers (>=4.13.2,<5.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)"
]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
quote-style = "single"

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev book_catalog/app.py'
test = 'pytest -s -x --cov=book_catalog -vv'
post_test = 'coverage html'
